Improve SNR calculation if neighbor channels are in use.

+++ src/tests/backend_tests.cpp
@@ -39,7 +39,7 @@
 
 class TestRadioInterface : public RadioControllerInterface {
     public:
-        virtual void onSNR(int snr) override { (void)snr; }
+        virtual void onSNR(float snr) override { (void)snr; }
         virtual void onFrequencyCorrectorChange(int fine, int coarse) override { (void)fine; (void)coarse; }
         virtual void onSyncChange(char isSync) override { (void)isSync; }
         virtual void onSignalPresence(bool isSignal) override { (void)isSignal; }
